+++ react-image-editor/src/App.css
@@ -57,3 +57,10 @@
 .downloadBtn {
   cursor: pointer;
 }
+.loginModal,
+.signUpModal{
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
